requirements/base.txt
# Django
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Celery
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# LLM
google-genai==0.3.0

# Image Processing
Pillow==10.2.0

# Storage
boto3==1.34.34

# Utils
python-dotenv==1.0.0
python-decouple==3.8

# Security
django-cors-headers==4.3.1

# API Documentation
drf-spectacular==0.27.1

# Monitoring
sentry-sdk==1.40.0
django-modeltranslation
